Convert punishment IDs to 64-bit and integrate Snowflake IDs.
Updated punishment IDs from 32-bit integers to 64-bit longs for scalability. Integrated the Snowflake ID generator to ensure globally unique identifiers. Adjusted database schemas, models, and associated logic accordingly.

+++ core/src/main/java/me/mykindos/betterpvp/core/client/punishments/PunishmentRepository.java
@@ -88,7 +88,7 @@ public class PunishmentRepository implements IRepository<Punishment> {
                     .fetch();
 
             for (var punishmentRecord : records) {
-                int punishmentId = punishmentRecord.get(PUNISHMENTS.ID);
+                long punishmentId = punishmentRecord.get(PUNISHMENTS.ID);
                 long punishedClientId = punishmentRecord.get(PUNISHMENTS.CLIENT);
                 UUID punishedClientUUID = UUID.fromString(punishmentRecord.get(CLIENTS.UUID));
                 IPunishmentType type = PunishmentTypes.getPunishmentType(punishmentRecord.get(PUNISHMENTS.TYPE));
